pyproject.toml
[project]
name = "fast-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "fastapi>=0.115.13",
    "google-api-python-client>=2.176.0",
    "langchain>=0.3.26",
    "langchain-mcp-adapters>=0.1.7",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.10",
    "mcp[cli]>=1.9.4",
    "numpy>=2.3.1",
    "ollama>=0.5.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.46.1",
    "tavily-python>=0.7.8",
    "ultralytics>=8.3.162",
    "uvicorn[standard]>=0.34.3",
    "youtube-transcript-api>=1.1.1",
]
